validation added for unauthorized user while api calling

+++ src/hooks/mutations/favorites.mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { getErrorMessage } from '@/src/utils/getErrorMessage';
 import actotaApi from '@/src/lib/apiClient';
-
+import { toast } from 'react-hot-toast';
 
 // Subscribe mutation
 const useAddFavorites = () => {
@@ -9,6 +9,11 @@ const useAddFavorites = () => {
     mutationFn: async (favoriteId: string) => {
       try {
         const user = JSON.parse(localStorage.getItem('user') || '{}');
+        console.log(user, "user");
+        if(!user.user_id){
+          toast.error("Please login to add favorites");
+          throw new Error("Please login to add favorites");
+        }
             const response = await actotaApi.post(
           `/api/account/${user?.user_id}/favorites/${favoriteId}`,
           {},
@@ -31,6 +36,10 @@ const useRemoveFavorites = () => {
       mutationFn: async (favoriteId: string) => {
         try {
           const user = JSON.parse(localStorage.getItem('user') || '{}');
+          if(!user.user_id){
+            toast.error("Please login to remove favorites");
+            throw new Error("Please login to remove favorites");
+          }
               const response = await actotaApi.delete(
             `/api/account/${user?.user_id}/favorites/${favoriteId}`,
